Update create_tables.sql copy source path

+++ upgrade_php.sh
@@ -299,7 +299,12 @@ if [[ "$UPGRADE_PMA" = "y" || "$UPGRADE_PMA" = "Y" ]];then
     fi
     mkdir -p /data/www/default/phpmyadmin/upload/
     mkdir -p /data/www/default/phpmyadmin/save/
-    cp -f /data/www/default/phpmyadmin/examples/create_tables.sql /data/www/default/phpmyadmin/upload/
+    if [ -s /data/www/default/phpmyadmin/examples/create_tables.sql ]; then
+        cp -f /data/www/default/phpmyadmin/examples/create_tables.sql /data/www/default/phpmyadmin/upload/
+    elif [ -s /data/www/default/phpmyadmin/sql/create_tables.sql ]; then
+        cp -f /data/www/default/phpmyadmin/sql/create_tables.sql /data/www/default/phpmyadmin/upload/
+    fi
+
     chown -R apache:apache /data/www/default/phpmyadmin
     # clean phpMyAdmin archive
     cd $cur_dir
